+++ frontend/.storybook/main.js
@@ -26,6 +26,8 @@ module.exports = {
       '@create-study-page': resolve(__dirname, '../src/pages/create-study-page'),
       '@my-study-page': resolve(__dirname, '../src/pages/my-study-page'),
       '@study-room-page': resolve(__dirname, '../src/pages/study-room-page'),
+      '@login-redirect-page': resolve(__dirname, '../src/pages/login-redirect-page'),
+      '@error-page': resolve(__dirname, '../src/pages/error-page'),
       '@layout': resolve(__dirname, '../src/layout'),
       '@hooks': resolve(__dirname, '../src/hooks'),
       '@mocks': resolve(__dirname, '../src/mocks'),
