+++ frontend/src/pages/index.ts
@@ -1,5 +1,5 @@
-export { default as ErrorPage } from '@pages/error-page/ErrorPage';
-export { default as LoginRedirectPage } from '@pages/login-redirect-page/LoginRedirectPage';
+export { default as ErrorPage } from '@error-page/ErrorPage';
+export { default as LoginRedirectPage } from '@login-redirect-page/LoginRedirectPage';
 export { default as MainPage } from '@main-page/MainPage';
 export { default as DetailPage } from '@detail-page/DetailPage';
 export { default as CreateStudyPage } from '@create-study-page/CreateStudyPage';
