+++ frontend/webpack/webpack.common.js
@@ -52,6 +52,8 @@ module.exports = {
       '@create-study-page': resolve(__dirname, '../src/pages/create-study-page'),
       '@my-study-page': resolve(__dirname, '../src/pages/my-study-page'),
       '@study-room-page': resolve(__dirname, '../src/pages/study-room-page'),
+      '@error-page': resolve(__dirname, '../src/pages/error-page'),
+      '@login-redirect-page': resolve(__dirname, '../src/pages/login-redirect-page'),
       '@layout': resolve(__dirname, '../src/layout'),
       '@hooks': resolve(__dirname, '../src/hooks'),
       '@mocks': resolve(__dirname, '../src/mocks'),
